Separate reduction code from _kernel.pyx
