+++ cupy/core/_routines_manipulation.pxd
@@ -10,7 +10,7 @@ cdef class broadcast:
         readonly Py_ssize_t size
         readonly Py_ssize_t nd
 
-cdef _broadcast_core(list arrays, vector.vector[Py_ssize_t]& shape)
+
 cdef _ndarray_shape_setter(ndarray self, newshape)
 cdef ndarray _ndarray_reshape(ndarray self, tuple shape, order)
 cdef ndarray _ndarray_transpose(ndarray self, tuple axes)
