+++ cupy/core/_routines_manipulation.pyx
@@ -8,76 +8,14 @@ from cupy.core import _errors
 from cupy.core._kernel import ElementwiseKernel
 from cupy.core._ufuncs import elementwise_copy
 
-cimport cython  # NOQA
 cimport cpython  # NOQA
+cimport cython  # NOQA
 from libcpp cimport vector
 
 from cupy.core cimport _routines_indexing as _indexing
 from cupy.core cimport core
 from cupy.core.core cimport ndarray
 from cupy.core cimport internal
-
-
-cdef Py_ssize_t PY_SSIZE_T_MAX = sys.maxsize
-
-
-cdef _broadcast_core(list arrays, vector.vector[Py_ssize_t]& shape):
-    cdef Py_ssize_t i, j, s, smin, smax, a_ndim, a_sh, nd
-    cdef vector.vector[Py_ssize_t] strides
-    cdef vector.vector[int] index
-    cdef ndarray a
-    cdef list ret
-
-    shape.clear()
-    index.reserve(len(arrays))
-    nd = 0
-    for i, x in enumerate(arrays):
-        if not isinstance(x, ndarray):
-            continue
-        a = x
-        index.push_back(i)
-        nd = max(nd, <Py_ssize_t>a._shape.size())
-
-    if index.size() == 0:
-        return
-
-    shape.reserve(nd)
-    for i in range(nd):
-        smin = PY_SSIZE_T_MAX
-        smax = 0
-        for j in index:
-            a = arrays[j]
-            a_ndim = <Py_ssize_t>a._shape.size()
-            if i >= nd - a_ndim:
-                s = a._shape[i - (nd - a_ndim)]
-                smin = min(smin, s)
-                smax = max(smax, s)
-        if smin == 0 and smax > 1:
-            raise ValueError(
-                'shape mismatch: objects cannot be broadcast to a '
-                'single shape')
-        shape.push_back(0 if smin == 0 else smax)
-
-    for i in index:
-        a = arrays[i]
-        if internal.vector_equal(a._shape, shape):
-            continue
-
-        strides.assign(nd, <Py_ssize_t>0)
-        a_ndim = <Py_ssize_t>a._shape.size()
-        for j in range(a_ndim):
-            a_sh = a._shape[j]
-            if a_sh == shape[j + nd - a_ndim]:
-                strides[j + nd - a_ndim] = a._strides[j]
-            elif a_sh != 1:
-                raise ValueError(
-                    'operands could not be broadcast together with shapes '
-                    '{}'.format(
-                        ', '.join([str(x.shape) if isinstance(x, ndarray)
-                                   else '()' for x in arrays])))
-
-        # TODO(niboshi): Confirm update_x_contiguity flags
-        arrays[i] = a._view(shape, strides, True, True)
 
 
 @cython.final
@@ -103,7 +41,7 @@ cdef class broadcast:
     def __init__(self, *arrays):
         cdef vector.vector[Py_ssize_t] shape
         cdef list val = list(arrays)
-        _broadcast_core(val, shape)
+        internal._broadcast_core(val, shape)
         self.values = tuple(val)
         self.shape = tuple(shape)
         self.nd = <Py_ssize_t>shape.size()
