+++ cupy/core/internal.pyx
@@ -5,10 +5,17 @@ cimport cython  # NOQA
 from libcpp cimport bool as cpp_bool
 from libc.stdint cimport uint32_t
 
+from cupy.core.core cimport ndarray
+
+import sys
+
 
 cdef extern from 'halffloat.h':
     uint16_t npy_floatbits_to_halfbits(uint32_t f)
     uint32_t npy_halfbits_to_floatbits(uint16_t h)
+
+
+cdef Py_ssize_t PY_SSIZE_T_MAX = sys.maxsize
 
 
 @cython.profile(False)
@@ -291,3 +298,62 @@ cdef inline int _normalize_order(order, cpp_bool allow_k=True) except? 0:
     else:
         raise TypeError('order not understood')
     return order_char
+
+
+cdef _broadcast_core(list arrays, vector.vector[Py_ssize_t]& shape):
+    cdef Py_ssize_t i, j, s, smin, smax, a_ndim, a_sh, nd
+    cdef vector.vector[Py_ssize_t] strides
+    cdef vector.vector[int] index
+    cdef ndarray a
+    cdef list ret
+
+    shape.clear()
+    index.reserve(len(arrays))
+    nd = 0
+    for i, x in enumerate(arrays):
+        if not isinstance(x, ndarray):
+            continue
+        a = x
+        index.push_back(i)
+        nd = max(nd, <Py_ssize_t>a._shape.size())
+
+    if index.size() == 0:
+        return
+
+    shape.reserve(nd)
+    for i in range(nd):
+        smin = PY_SSIZE_T_MAX
+        smax = 0
+        for j in index:
+            a = arrays[j]
+            a_ndim = <Py_ssize_t>a._shape.size()
+            if i >= nd - a_ndim:
+                s = a._shape[i - (nd - a_ndim)]
+                smin = min(smin, s)
+                smax = max(smax, s)
+        if smin == 0 and smax > 1:
+            raise ValueError(
+                'shape mismatch: objects cannot be broadcast to a '
+                'single shape')
+        shape.push_back(0 if smin == 0 else smax)
+
+    for i in index:
+        a = arrays[i]
+        if vector_equal(a._shape, shape):
+            continue
+
+        strides.assign(nd, <Py_ssize_t>0)
+        a_ndim = <Py_ssize_t>a._shape.size()
+        for j in range(a_ndim):
+            a_sh = a._shape[j]
+            if a_sh == shape[j + nd - a_ndim]:
+                strides[j + nd - a_ndim] = a._strides[j]
+            elif a_sh != 1:
+                raise ValueError(
+                    'operands could not be broadcast together with shapes '
+                    '{}'.format(
+                        ', '.join([str(x.shape) if isinstance(x, ndarray)
+                                   else '()' for x in arrays])))
+
+        # TODO(niboshi): Confirm update_x_contiguity flags
+        arrays[i] = a._view(shape, strides, True, True)
